# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SOWISO Etiquette API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /culture/origins:
    get:
      operationId: getCulturalOrigins
      tags: [culture]
      summary: Get historical and cultural origin records for a region
      parameters:
        - name: region_code
          in: query
          required: true
          schema:
            type: string
          description: ISO country code (e.g. GB, CN, CA)
        - name: domain
          in: query
          required: false
          schema:
            type: string
            enum: [dining, business, greetings, gift-giving, dress]
          description: Filter by domain. Omit for all domains.
      responses:
        "200":
          description: Cultural origin records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CulturalOrigin"
        "400":
          description: Missing or invalid region code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /culture/protocols:
    get:
      operationId: getCultureProtocols
      tags: [culture]
      summary: Get culture protocols for a region and pillar
      parameters:
        - name: region_code
          in: query
          required: true
          schema:
            type: string
          description: ISO country code (e.g. GB, CN, CA)
        - name: pillar
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5
          description: Pillar number (1-5). Omit for all pillars.
        - name: context
          in: query
          required: false
          schema:
            type: string
            enum: [business, social, general]
        - name: pillar_code
          in: query
          required: false
          schema:
            type: string
            enum: [Z1, Z2, Z3, Z4, Z5]
          description: Filter to verified CC Screening records under a single 5-pillar code (Z1-Z5).
        - name: verified_only
          in: query
          required: false
          schema:
            type: boolean
          description: When true, only verified CC Screening records (verified = true, source_book set) are returned.
        - name: locale
          in: query
          required: false
          schema:
            type: string
          description: BCP 47 locale; resolves rule_cc_i18n into display_rule.
        - name: situational_interests
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated sphere keys (e.g. business,gastronomy) used to sort matching contexts to the top.
      responses:
        "200":
          description: Culture protocols
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CultureProtocol"
  /culture/compass:
    get:
      operationId: getCultureCompass
      tags: [culture]
      summary: Get cultural compass summary for all available regions
      parameters:
        - name: locale
          in: query
          required: false
          schema:
            type: string
          description: "BCP 47 locale code (e.g. en-GB, en-US, nl-NL, fr-FR). Defaults to en-GB."
      responses:
        "200":
          description: Cultural compass data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CultureCompassEntry"
  /culture/compass/{regionCode}:
    get:
      operationId: getCultureCompassRegion
      tags: [culture]
      summary: Get detailed cultural compass for a specific region
      parameters:
        - name: regionCode
          in: path
          required: true
          schema:
            type: string
        - name: locale
          in: query
          required: false
          schema:
            type: string
          description: "BCP 47 locale code (e.g. en-GB, en-US, nl-NL, fr-FR). Defaults to en-GB."
      responses:
        "200":
          description: Cultural compass detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CultureCompassDetail"
        "404":
          description: Region not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CulturalOrigin:
      type: object
      properties:
        id:
          type: integer
        region_code:
          type: string
        domain:
          type: string
        tradition:
          type: string
        origin_summary:
          type: string
        era:
          type: string
        influences:
          type: array
          items:
            type: string
        connected_rule:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - region_code
        - domain
        - tradition
        - origin_summary
        - era
        - influences
        - connected_rule
    CultureProtocol:
      type: object
      properties:
        id:
          type: integer
        region_code:
          type: string
        pillar:
          type: integer
        rule_type:
          type: string
        rule_description:
          type: string
        gender_applicability:
          type: string
          enum: [all, masculine, feminine, fluid]
        context:
          type: string
          enum: [business, social, general, formal, dining]
        display_rule:
          type: string
          description: Locale-aware display text. Falls back to rule_cc, then rule_description.
        pillar_code:
          type: string
          enum: [Z1, Z2, Z3, Z4, Z5]
          nullable: true
          description: CC Screening Worker 5-pillar code.
        subcategory:
          type: string
          nullable: true
        rule_cc:
          type: string
          nullable: true
          description: C&C mentor-formulation, used as the display rule when present.
        urgency:
          type: integer
          minimum: 1
          maximum: 3
          nullable: true
          description: 1 = nice-to-know, 2 = important, 3 = critical (First Aid).
        verified:
          type: boolean
          nullable: true
        source_book:
          type: string
          nullable: true
      required:
        - id
        - region_code
        - pillar
        - rule_type
        - rule_description
        - gender_applicability
        - context
    CultureCompassEntry:
      type: object
      properties:
        region_code:
          type: string
        region_name:
          type: string
        core_value:
          type: string
        biggest_taboo:
          type: string
        flag_emoji:
          type: string
        has_content:
          type: boolean
          description: "True when this region has at least one locale of compass content ready."
        sphere_highlights:
          type: array
          description: "Situational interest spheres (e.g. business, leisure) for which this region currently has dedicated highlight content."
          items:
            type: string
      required:
        - region_code
        - region_name
        - core_value
        - biggest_taboo
        - flag_emoji
        - has_content
    CultureCompassDetail:
      type: object
      properties:
        region_code:
          type: string
        region_name:
          type: string
        flag_emoji:
          type: string
        core_value:
          type: string
        biggest_taboo:
          type: string
        dining_etiquette:
          type: string
        language_notes:
          type: string
        gift_protocol:
          type: string
        dress_code:
          type: string
        dos:
          type: array
          items:
            type: string
        donts:
          type: array
          items:
            type: string
        mehrabian_weight:
          type: object
          properties:
            nonverbal:
              type: number
            tone:
              type: number
            words:
              type: number
            note:
              type: string
          required:
            - nonverbal
            - tone
            - words
        sphere_highlights:
          type: array
          description: "Compass field names that include sphere-specific highlight content for the requested situational interests (e.g. dining_etiquette, dress_code)."
          items:
            type: string
      required:
        - region_code
        - region_name
        - flag_emoji
        - core_value
        - biggest_taboo
        - dining_etiquette
        - language_notes
        - gift_protocol
        - dress_code
        - dos
        - donts
...